Added loading game

+++ sem/src/main/java/cz/cvut/fel/pjv/semestral/kadardmy/Utilities/Timer.java
@@ -30,7 +30,7 @@ public class Timer {
     private long enabledStartTime = 0;
 
 
-    public Timer(Board.Color color, int time) {
+    public Timer(Board.Color color, long time) {
         this.time = time;
         this.color = color;
     }
